+++ Ficha.java
@@ -7,10 +7,10 @@ public class Ficha {
         color = pcolor;
     }
     public void avanzar(){
-        double pasos = dadoActu.lanzar();
+        int pasos = dadoActu.lanzar();
         posicion += pasos;
         System.out.println("Es el turno de la ficha: " + color + "\n");
-        System.out.println("¡La ficha: " + color + " avanza " + (int)pasos + " casillas!\n");
+        System.out.println("¡La ficha: " + color + " avanza " + pasos + " casillas!\n");
         if (posicion >= 40)
             posicion = 40;
         System.out.println("La casilla actual de la ficha es: " + posicion + "\n");
